fix: resolve hostname
Fixes issue where the hostname only resolves into an IP address when the
sink is created. The very nature of hostnames is to provide a route to
instances whose IP address may change, and these code changes allows for
this situation.

One limitation with the changes is that currently only IPv4 addresses are
supported, but an upcoming PR will solve that issue.

Closes #29